+++ ios/sourcemaps.sh
@@ -51,12 +51,14 @@ fi
   local inferred_code=$(/usr/libexec/PlistBuddy -c 'print CFBundleVersion' "$PROJECT_DIR/$INFOPLIST_FILE")
   local version_code=$(resolve_var "Version Code" "INSTABUG_APP_VERSION_CODE" "$inferred_code" | tail -n 1)
 
+if [ -n "$sourcemap_file" ]; then
   node $instabug_dir/bin/index.js upload-sourcemaps \
       --platform ios \
       --file $sourcemap_file \
       --token $app_token \
       --name $version_name \
       --code $version_code
+      fi
 }
 
 generate_sourcemaps() {